+++ entrypoint.sh
@@ -5,7 +5,7 @@ source ${INVOICEPLANE_RUNTIME_DIR}/functions
 [[ $DEBUG == true ]] && set -x
 
 case ${1} in
-  app:invoiceplane|app:nginx|app:backup:create)
+  app:invoiceplane|app:nginx|app:backup:create|app:backup:restore)
 
     initialize_system
 
@@ -24,6 +24,10 @@ case ${1} in
         shift 1
         backup_create
         ;;
+      app:backup:restore)
+        shift 1
+        backup_restore $@
+        ;;
     esac
     ;;
   app:help)
@@ -31,6 +35,7 @@ case ${1} in
     echo " app:invoiceplane     - Starts the InvoicePlane php5-fpm server (default)"
     echo " app:nginx            - Starts the nginx server"
     echo " app:backup:create    - Create a backup"
+    echo " app:backup:restore   - Restore an existing backup"
     echo " app:help             - Displays the help"
     echo " [command]            - Execute the specified command, eg. bash."
     ;;